t820: add initial kickstart-oms.py
even with its flaws, this is some serious awesome!

overall, this script completes three primary tasks:

  1) install and configure salt-minion on the host
  2) run state.high to apply some initial states
  3) run state.highstate to apply more advanced states

default config and state data is included in the script, so no
external config is necessary and the script can run by itself.

when run this way, kickstart-oms.py will apply a set of base
states that install minimal packages needed to clone the
oms-deploy repo and setup the host for a run of state.highstate.

state.highstate will update the system using the salt states and
pillar definitions from the oms-deploy repo.

kickstart-oms.py may be run with the following flags:

  -d  enables debug logging to share some helpful details
  -t  enables test (noop) mode, no actions will be taken
  -H  runs state.highstate after the host is setup with the
      base states
  -h  prints help messages
  -c  will let you provide a list of YAML-formatted configs

lastly, kickstart-oms.py will accept external configuration to
override the defaults embedded in the script. this is where
we get super awesome, as we are free to..

    A) provide any minion config we wish
    B) include any base states to apply

thus.. this script may be used to kickstart a host with an
arbitrary set of base states, then run state.highstate to
apply a second round of updates, allowing us to do pretty much
anything we want with salt.