t1906: kickstart kickstart!
OMS the holonic multi-verse!

* remove the old kickstart-oms.sh
* update .gitignore to ignore SSH keys as id_rsa*
* include a README with instructions RE SSH keys in config/keys/
* include a script to kickstart the kickstart process on a remote host
* include a script to wrap up running kickstart in a tmux session (kick-tmux.sh)
* include a script to wrap up running kickstart itself (run-kickstart.sh)
* make kickstart-oms.py executable as part of this (we rely on this)

This is intended to be run from a developer's primary development host, or
similar central, command and control host.

The purpose is to take a remote host through the process to setup the scripts
and configuration to run the OMS kickstart process.

If you want to edit the actual kickstart incantation (which config to use, etc)
edit run-kickstart.sh