t1908: add Dockerfiles and docs to kickstart docker containers
This is the result of some tests I've run - using docker as an alternative
to full VMs became a necessity (to setup and use) in oms-kickstart's
development. Using a VM to test updates to oms-kickstart became too costly,
though this is not to say docker is fantastic - it's a beast in its own
right.

This is nothing extrodinary or extensive, just some simple resources to get
started with docker and OMS - include scripts to build ubuntu images from
scratch, as well as the Dockerfiles to build an `oms/base` and `oms/kick`
image.

The OMS TCF has not been extensively tested, though it will need updates to
maintain peace with docker - this commit is just the generic structures
needed to build two types of docker images, using oms-kickstart to manage the
contents of the image.