t1982: use embedded pillar to seed reclass' master_tops
the master tops is the list of states to apply to a node - this embedded
pillar dictionary (the classes list, specifically), is used to seed the
node map used by reclass.

note that this list is in addition to the base class. at present the list
includes only one item, the umbrella state, `fullstack`. this will change
in future releases.